[Refactor]-useImageUploader 훅 생성, 이미지 업로드 관련 폴더 구조 개선
