+++ src/lib/common/api.ts
@@ -1,22 +1,10 @@
 import { storage } from "@/firebase";
-import imageCompression from "browser-image-compression";
 import {
   deleteObject,
   getDownloadURL,
   ref,
   uploadBytes,
 } from "firebase/storage";
-
-const imageCompressionOptions = {
-  maxSizeMB: 1,
-  maxWidthOrHeight: 800,
-  useWebWorker: true,
-  fileType: "image/webp",
-};
-
-export const compressAndConvertImage = async (file: File): Promise<Blob> => {
-  return await imageCompression(file, imageCompressionOptions);
-};
 
 export const uploadImageToFirebase = async (
   file: Blob,
@@ -27,7 +15,7 @@ export const uploadImageToFirebase = async (
   return await getDownloadURL(storageRef);
 };
 
-export const deleteImageFormFirebaes = async (url: string): Promise<void> => {
+export const deleteImageFromFirebase = async (url: string): Promise<void> => {
   const imageRef = ref(storage, url);
   try {
     await deleteObject(imageRef);
